Revert removed comma
While not required, because the sentence is compound, one should
use a comma in order to offset "such as NumPy" to mirror the pause
that someone would use if reading the text out loud. Without the
comma, the comma after NumPy should also be removed. But if this is
done, then the sentence, when spoken, sounds like a run-on sentence.