Fix design principle description
APIs, such as unique and friends, are not "optional extensions" as
they are defined. Instead, they are simply optional as part of the
main spec. NumPy, for example, can support `unique` and friends
and must in order to be specification-compliant. NumPy could,
however, choose not to implement FFT APIs, the set of which is
an optional extension.