Fix use of "combination"
When generating signatures, we did not generate combinations.
Instead, we generated signatures based on input argument and
associated type pairs. For example, if a unary function was provided
a list, we generated a signature with the first argument was a list.
Use of the term "combinations" might imply that, for a binary function,
we generated signatures based on the sets of respective types. This
would not be true.