Remove downstream status updates
Support in scikit-learn and SciPy could change. Including these
asides does not add much to the discussion, especially SciPy
which is still not guaranteed to see the light of day. The discussion
is focused on perf implications. Whether changes are adopted by
downstream libraries is not the point in this context.